refactor: load partners data from supabase
