--- core/partners/store.ts
@@ -1,358 +1,0 @@
-import type { Partner } from "./types";
-
-const partners: Partner[] = [
-  {
-    id: "stride-coalition",
-    name: "Stride Business Coalition",
-    organizationType: "Chamber of Commerce",
-    focus: "Local founders & growth-stage businesses",
-    city: "Austin",
-    state: "TX",
-    stage: "Discovery",
-    owner: "Jordan Blake",
-    model: "Referral Exchange",
-    potentialValue: 185000,
-    warmIntroductions: 9,
-    mutualClients: 6,
-    readinessScore: 78,
-    notes: [
-      "Looking for partners to co-host quarterly growth roundtables.",
-      "Interested in revenue operations workshops for member companies.",
-    ],
-    website: "https://stridecoalition.example.com",
-    lastInteraction: "2025-10-07",
-    ecosystemFit: "Anchor",
-    strengths: ["Trusted local network", "Strong executive attendance"],
-    needs: ["Modern enablement content", "Intro scripts"],
-    contacts: [
-      {
-        id: "stride-1",
-        name: "Alexis Moore",
-        role: "Executive Director",
-        email: "alexis@stridecoalition.org",
-        phone: "512-555-1023",
-      },
-      {
-        id: "stride-2",
-        name: "Marco Alvarez",
-        role: "Member Success Lead",
-        email: "marco@stridecoalition.org",
-      },
-    ],
-    opportunities: [
-      {
-        id: "stride-op-1",
-        name: "Q4 Founders Exchange",
-        type: "Event",
-        value: 45000,
-        status: "In Motion",
-        targetClient: "Regional SaaS founders",
-        expectedClose: "2025-11-02",
-      },
-      {
-        id: "stride-op-2",
-        name: "Member Referral Program",
-        type: "Referral",
-        value: 90000,
-        status: "Introduced",
-        targetClient: "Growth-stage operators",
-        expectedClose: "2025-12-01",
-      },
-    ],
-    initiatives: [
-      {
-        id: "stride-init-1",
-        title: "Local Expansion Playbook",
-        owner: "Jordan Blake",
-        status: "Active",
-        dueDate: "2025-11-15",
-        description: "Co-develop a playbook for member companies expanding sales coverage.",
-      },
-      {
-        id: "stride-init-2",
-        title: "Community Spotlight Series",
-        owner: "Alexis Moore",
-        status: "Planning",
-        dueDate: "2025-12-05",
-        description: "Highlight mutual success stories to drive warm introductions.",
-      },
-    ],
-    interactions: [
-      {
-        id: "stride-touch-1",
-        date: "2025-10-07",
-        type: "Meeting",
-        summary: "Reviewed referral workflow and agreed on co-branded assets.",
-        nextStep: "Send pilot asset package for review by 10/12.",
-        sentiment: "Positive",
-      },
-      {
-        id: "stride-touch-2",
-        date: "2025-09-24",
-        type: "Intro",
-        summary: "Introduced to two member CROs evaluating revenue operations support.",
-        nextStep: "Follow up with CROs week of 10/14.",
-        sentiment: "Positive",
-      },
-    ],
-    resources: [
-      {
-        id: "stride-res-1",
-        title: "Referral Intro Script",
-        type: "Playbook",
-        url: "https://assets.trs.local/stride-referral.pdf",
-        notes: "Customize with coalition branding before distribution.",
-      },
-      {
-        id: "stride-res-2",
-        title: "Co-Hosted Event Brief",
-        type: "Deck",
-        url: "https://assets.trs.local/stride-event-brief.pdf",
-      },
-    ],
-  },
-  {
-    id: "metro-innovation",
-    name: "Metro Innovation Hub",
-    organizationType: "Economic Development",
-    focus: "Emerging tech companies",
-    city: "Dallas",
-    state: "TX",
-    stage: "Initial Outreach",
-    owner: "Jordan Blake",
-    model: "Co-Marketing",
-    potentialValue: 98000,
-    warmIntroductions: 4,
-    mutualClients: 2,
-    readinessScore: 61,
-    notes: [
-      "Interested in a co-hosted pipeline diagnostics clinic.",
-      "Needs collateral tailored to public-private partnerships.",
-    ],
-    website: "https://metroinnovation.example.com",
-    lastInteraction: "2025-10-03",
-    ecosystemFit: "Strategic",
-    strengths: ["Strong municipal backing", "Marketing reach"],
-    needs: ["Proof points", "Executive sponsor"],
-    contacts: [
-      {
-        id: "metro-1",
-        name: "Priya Raman",
-        role: "Program Director",
-        email: "priya@metroinnovation.org",
-        phone: "214-555-4673",
-      },
-    ],
-    opportunities: [
-      {
-        id: "metro-op-1",
-        name: "Pipeline Diagnostics Clinic",
-        type: "Event",
-        value: 35000,
-        status: "Sourcing",
-        targetClient: "Incubator cohort",
-        expectedClose: "2025-11-20",
-      },
-    ],
-    initiatives: [
-      {
-        id: "metro-init-1",
-        title: "Startup Week Programming",
-        owner: "Priya Raman",
-        status: "Planning",
-        dueDate: "2025-11-28",
-        description: "Design a breakout session on revenue readiness for founders.",
-      },
-    ],
-    interactions: [
-      {
-        id: "metro-touch-1",
-        date: "2025-10-03",
-        type: "Call",
-        summary: "Introductory call focused on community impact metrics.",
-        nextStep: "Share examples of impact reporting.",
-        sentiment: "Positive",
-      },
-    ],
-    resources: [
-      {
-        id: "metro-res-1",
-        title: "Community Impact One-Pager",
-        type: "One-Pager",
-        url: "https://assets.trs.local/community-impact.pdf",
-      },
-    ],
-  },
-  {
-    id: "artisan-alliance",
-    name: "Artisan Retail Alliance",
-    organizationType: "Business Association",
-    focus: "Boutique retail and experiential brands",
-    city: "San Antonio",
-    state: "TX",
-    stage: "Pilot Collaboration",
-    owner: "Morgan Lee",
-    model: "Referral Exchange",
-    potentialValue: 132000,
-    warmIntroductions: 6,
-    mutualClients: 3,
-    readinessScore: 84,
-    notes: [
-      "Excited about mutual gift-card program.",
-      "Needs enablement for member success teams.",
-    ],
-    website: "https://artisanalliance.example.com",
-    lastInteraction: "2025-10-08",
-    ecosystemFit: "Strategic",
-    strengths: ["High-trust merchant community", "Existing referral cadence"],
-    needs: ["Automation", "Shared success stories"],
-    contacts: [
-      {
-        id: "artisan-1",
-        name: "Taylor Nguyen",
-        role: "Director of Partnerships",
-        email: "taylor@artisanalliance.org",
-      },
-      {
-        id: "artisan-2",
-        name: "Jamie Ortiz",
-        role: "Experience Lead",
-        email: "jamie@artisanalliance.org",
-      },
-    ],
-    opportunities: [
-      {
-        id: "artisan-op-1",
-        name: "Mutual Gift Card Program",
-        type: "Joint Project",
-        value: 72000,
-        status: "In Motion",
-        targetClient: "Alliance member merchants",
-        expectedClose: "2025-10-31",
-      },
-      {
-        id: "artisan-op-2",
-        name: "Holiday Retail Tour",
-        type: "Event",
-        value: 60000,
-        status: "Introduced",
-        targetClient: "Retail foot-traffic initiatives",
-        expectedClose: "2025-11-18",
-      },
-    ],
-    initiatives: [
-      {
-        id: "artisan-init-1",
-        title: "Member Enablement Sprint",
-        owner: "Morgan Lee",
-        status: "Active",
-        dueDate: "2025-10-25",
-        description: "Deliver enablement kit and shared success metrics dashboard.",
-      },
-    ],
-    interactions: [
-      {
-        id: "artisan-touch-1",
-        date: "2025-10-08",
-        type: "Meeting",
-        summary: "Joint roadmap workshop finalizing Q4 initiatives.",
-        nextStep: "Send revised roadmap deck by 10/10.",
-        sentiment: "Positive",
-      },
-      {
-        id: "artisan-touch-2",
-        date: "2025-09-19",
-        type: "Event",
-        summary: "Co-hosted member dinner; discussed integrated referral flows.",
-        sentiment: "Positive",
-      },
-    ],
-    resources: [
-      {
-        id: "artisan-res-1",
-        title: "Alliance Enablement Kit",
-        type: "Checklist",
-        url: "https://assets.trs.local/alliance-enablement.pdf",
-      },
-    ],
-  },
-  {
-    id: "greenway-wellness",
-    name: "Greenway Wellness Collective",
-    organizationType: "Wellness Cooperative",
-    focus: "Health & wellness providers",
-    city: "Houston",
-    state: "TX",
-    stage: "Discovery",
-    owner: "Morgan Lee",
-    model: "Community",
-    potentialValue: 86000,
-    warmIntroductions: 5,
-    mutualClients: 4,
-    readinessScore: 69,
-    notes: [
-      "Requires HIPAA-conscious messaging.",
-      "Exploring co-branded wellness workshops.",
-    ],
-    website: "https://greenwaywellness.example.com",
-    lastInteraction: "2025-10-04",
-    ecosystemFit: "Emerging",
-    strengths: ["Dedicated community team", "Recurring events calendar"],
-    needs: ["Trusted compliance guidance", "Wellness-focused case studies"],
-    contacts: [
-      {
-        id: "greenway-1",
-        name: "Riley Chen",
-        role: "Community Director",
-        email: "riley@greenwaywellness.org",
-      },
-    ],
-    opportunities: [
-      {
-        id: "greenway-op-1",
-        name: "Wellness Workshop Series",
-        type: "Event",
-        value: 42000,
-        status: "Introduced",
-        targetClient: "Collective members",
-        expectedClose: "2025-11-12",
-      },
-    ],
-    initiatives: [
-      {
-        id: "greenway-init-1",
-        title: "Compliance Content Refresh",
-        owner: "Morgan Lee",
-        status: "Planning",
-        dueDate: "2025-11-05",
-        description: "Adapt RevOS assets for wellness regulatory context.",
-      },
-    ],
-    interactions: [
-      {
-        id: "greenway-touch-1",
-        date: "2025-10-04",
-        type: "Call",
-        summary: "Aligned on referral criteria and warm intro process.",
-        nextStep: "Draft compliance-ready referral templates.",
-        sentiment: "Neutral",
-      },
-    ],
-    resources: [
-      {
-        id: "greenway-res-1",
-        title: "HIPAA Messaging Guide",
-        type: "Playbook",
-        url: "https://assets.trs.local/hipaa-messaging.pdf",
-      },
-    ],
-  },
-];
-
-export function getPartners(): Partner[] {
-  return partners;
-}
-
-export function getPartner(id: string): Partner | undefined {
-  return partners.find((partner) => partner.id === id);
-}
